+++ src/components/AppBar/AppBar.jsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import { Box, Container, Typography } from "@mui/material";
-import ModeSelect from "~/components/ModeSelect";
+import ModeSelect from "~/components/ModeSelect/ModeSelect";
 import AppsIcon from "@mui/icons-material/Apps";
 import TrelloIcon from "~/assets/trello.svg?react";
 import SvgIcon from "@mui/material/SvgIcon";
