+++ src/pages/Boards/BoardBar/Boardbar.jsx
@@ -36,7 +36,7 @@ function Boardbar() {
         overflowX: "auto",
         bgcolor: (theme) =>
           theme.palette.mode === "dark" ? "#34495e" : "#1976d2",
-        borderBottom: "1px solid #00bfa5",
+
         paddingX: 2,
       }}
     >
