+++ src/pages/Boards/BoardContent/ListColumns/Column/ListCards/Card/Card.jsx
@@ -1,0 +1,63 @@
+import React from "react";
+import Button from "@mui/material/Button";
+import { Card as MuiCard } from "@mui/material";
+import CardActions from "@mui/material/CardActions";
+import CardContent from "@mui/material/CardContent";
+import CardMedia from "@mui/material/CardMedia";
+import GroupIcon from "@mui/icons-material/Group";
+import CommentIcon from "@mui/icons-material/Comment";
+import AttachmentIcon from "@mui/icons-material/Attachment";
+import theme from "~/theme";
+import Paper from "@mui/material/Paper";
+import MenuList from "@mui/material/MenuList";
+import { Box, Container, Typography } from "@mui/material";
+function Card({ temporaryHidenMedia }) {
+  if (temporaryHidenMedia) {
+    return (
+      <MuiCard
+        sx={{
+          cursor: "pointer",
+          boxShadow: "0 1px 1px rgba(0,0,0,0.2)",
+          overflow: "unset",
+        }}
+      >
+        <CardContent sx={{ p: 1.5, "&:last-child": { p: 1.5 } }}>
+          <Typography>ThanhNgoDEV</Typography>
+        </CardContent>
+      </MuiCard>
+    );
+  }
+  return (
+    <MuiCard
+      sx={{
+        cursor: "pointer",
+        boxShadow: "0 1px 1px rgba(0,0,0,0.2)",
+        overflow: "unset",
+      }}
+    >
+      <CardMedia
+        sx={{ height: 140 }}
+        image="https://images2.thanhnien.vn/Uploaded/phucndh/2023_01_15/z4040523388756-c00d06cee10967450c92ab5ef4b703af-8795.jpg"
+        title="green iguana"
+      />
+      <CardContent sx={{ p: 1.5, "&:last-child": { p: 1.5 } }}>
+        <Typography>ThanhNgoDEV</Typography>
+      </CardContent>
+      <CardActions sx={{ p: "0 4px 8px 4px" }}>
+        <Button size="small" startIcon={<GroupIcon />}>
+          20
+        </Button>
+
+        <Button size="small" startIcon={<CommentIcon />}>
+          15
+        </Button>
+
+        <Button size="small" startIcon={<AttachmentIcon />}>
+          10
+        </Button>
+      </CardActions>
+    </MuiCard>
+  );
+}
+
+export default Card;
