+++ src/pages/Boards/BoardContent/ListColumns/Column/ListCards/ListCard.jsx
@@ -1,0 +1,35 @@
+import React from "react";
+import { Box, Container, Typography } from "@mui/material";
+import Card from "./Card/Card";
+
+function ListCard() {
+  return (
+    <Box
+      sx={{
+        display: "flex",
+        flexDirection: "column",
+        gap: 1,
+        p: "0 5px",
+        m: "0 5px",
+        overflowx: "hidden",
+        overflowY: "auto",
+        maxHeight: (theme) =>
+          `calc(${theme.trello.boardContentHeight} - ${theme.spacing(5)} - ${
+            theme.trello.columnHeaderHeight
+          } - ${theme.trello.columnFooterHeight})`,
+        "&::-webkit-scrollbar-thumb": {
+          backgroundColor: "#ced0da",
+          borderRadius: "8px",
+        },
+        "&::-webkit-scrollbar-thumb:hover": {
+          backgroundColor: "#bfc2cf",
+        },
+      }}
+    >
+      <Card />
+      <Card temporaryHidenMedia />
+    </Box>
+  );
+}
+
+export default ListCard;
