+++ src/pages/Boards/BoardContent/ListColumns/ListCloumn.jsx
@@ -1,0 +1,50 @@
+import { Box, Button, Container, Typography } from "@mui/material";
+import React from "react";
+import Column from "./Column/Column";
+import NoteAddIcon from "@mui/icons-material/NoteAdd";
+function ListCloumn() {
+  return (
+    <Box
+      sx={{
+        bgcolor: "inherit",
+        width: "100%",
+        height: "100%",
+        display: "flex",
+        overflowX: "auto",
+        overflowY: "hidden",
+        "&::-webkit-scrollbar-track": {
+          m: 2,
+        },
+      }}
+    >
+      <Column />
+      <Column />
+      {/* add new cloumn */}
+      <Box
+        sx={{
+          minWidth: "200px",
+          maxWidth: "200px",
+          mx: 2,
+          borderRadius: "6px",
+          height: "fit-content",
+          bgcolor: "#ffffff3d",
+        }}
+      >
+        <Button
+          startIcon={<NoteAddIcon />}
+          sx={{
+            color: "white",
+            width: "100%",
+            justifyContent: "flex-start",
+            pl: 2.5,
+            py: 1,
+          }}
+        >
+          Add new column
+        </Button>
+      </Box>
+    </Box>
+  );
+}
+
+export default ListCloumn;
